misc: fix -Wconversion warnings
Use type cast to fix -Wconversion warnings.

+++ src/mpid/common/bc/mpidu_bc.c
@@ -71,7 +71,7 @@ int MPIDU_bc_allgather(MPIR_Comm * allgather_comm, void *bc, int bc_len, int sam
     /* rank -> consecutive rank */
     for (i = 0; i < size; i++) {
         int node_id = MPIR_Process.node_map[i];
-        rank_map[i] = recv_offs[node_id];
+        rank_map[i] = (int) recv_offs[node_id];
         recv_offs[node_id]++;
     }
     /* shift offset back */
